macosx: add build schemes to Xcode project for complete compilation
This adds several build targets to the project for doing all needed
steps for compilation.

Additionally, they are some schemes included which are intended
to be used for compilation:
- "Rebuild everything" builds all targets in the right order, including
  extras/tools, contribs, bootstrap, configure, make and make VLC-dev.

  This scheme should be used once to setup everything. For later
  builds, you should switch to the Make project scheme.
- "Make project" runs "make && make VLC-dev.app", to only recompile
  files if needed. Additionally, it creates a VLC-dev.app for testing.

The end result will be stored in the build folder set in Xcode. Both
VLC.app and VLC-dev.app can be accessed by using the "Build artifacts"
folder, and showing the files in finder.