mmal/deinterlace: Rework to use pictures from vout pool
The reason for having a local picture pool within the deinterlace filter has
been that the mmal core did not allow to initialise the image_fx filter with
less than 20 output buffers. This limitation was removed from the mmal core,
so that it is possible to operate with as less as 2 output buffers, which
makes it possible to use pictures acquired from the core video output pool.
Besides being much more conformant to VLC core design this also avoids some
possible issues with buffer ownership and component lifecycle of the mmal
components.

Signed-off-by: Julian Scheel <julian@jusst.de>
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>