macosx: fix window ordering during startup (close #14486)
According to docu, application:willFinishLaunching: will be called
as one of the first steps of [NSApp run], before the main loop
will be actually started. Thus, it should make no performance
difference if we load the main window already in openIntf().

For some unknown reasons, a direct call to makeKeyAndOrderFront
works in run(), but not in app:willFinishLaunching:, though.