macosx: Split GoToTime panel into separate xib and window controller
This improves separation between main menu and the unrelated goto
time window.

Besides that, note that each time bindings are used with the file
owner, this owner needs to be an NSWindowController class. Only in
this case retain cycles are avoided (bindings have a strong
reference to the object used).

Thus, this commit helps to properly deinitialize the VLCMain object.