macosx: manage main window with an NSWindowController, simplify VLCMain
The window controller ensures that all top-level objects of the
mainwindow xib are properly released.

This removes the last occurence of VLCMain in xib files, thus
awakeFromNib will not be called for this object anymore. The code
has been moved to initThis removes the last occurence of VLCMain in xib files, thus
awakeFromNib will not be called for this object anymore. The code
has been moved to init.