avcodec: always emulate edges
The performance gains from not emulating edges was low if any, if not
negative. In any case, this was for non-DR cases only, so so much for
performance...

Also do not bother setting the flag at all with newer versions as it
was deprecated, and the underlying feature is always on.