macosx: Remove wrong "Change" button behavior in simple shortcut prefs
Remove the primary state of the "Change" button in the simple preferences
shotcut view, as the window already has a primary button, which is the one
to apply the settings. Having two buttons as primary buttons is bad UX
and very confusing as it's unclear which button will be triggered when
hitting the enter key.
Given that the "Change" button was not triggered by the enter key anyway,
this should not cause any app behavior changes.

Signed-off-by: Felix Paul Kühne <fkuehne@videolan.org>