macosx: Rewrote the About window auto-scroll behavior
It was impossible to scroll in the About Scroll Views as the code
was doing automatic scrolling. This patch removes all the related
code and introduces a new NSClipView subclass that handles the
automatic scrolling. If it detects that the user scrolls, the
animation is stopped and the user can scroll it as he wishes.

Signed-off-by: Felix Paul Kühne <fkuehne@videolan.org>