contrib/growl: Add growl patch that comments out logging
The patch comments out logging which causes a warning log message
when setting the growl bridge delegate to nil. This is a problem,
as we need to set it to nil, so that the growl bridge will release
our delegate. Else it will never be released and dealloc will
not get called.

Signed-off-by: Felix Paul Kühne <fkuehne@videolan.org>