input: use DEMUX_CAN_SEEK instead of STREAM_CAN_SEEK
Some demuxers support seeking even if the underlying byte stream is not
seekable (e.g. subtitles). Also some demuxers do not support seeking
regardless of the underlying byte stream.

So it makes more sense to check the status from the demuxer. Most
demuxers just copy or forward the value from the byte stream.

Note: if in doubt, a demuxer can always claim seek support, and
eventually fail both DEMUX_SET_TIME and DEMUX_SET_POSITION. The seek is
only used as a hint for user interfaces; it has no effects on the input
thread as such.