modules: add SoX Resampler audio_filter
This resampler supports 5 quality presets: from 0 to 4, 2 is the default and
correspond to "mq". It supports integer and float samples. Performances are way
better when this module is used as an "audio converter" (fixed sample rate).

See http://lastique.github.io/src_test/ for comparison with speexdsp.

It is deactivated for now.