input: check run-time against demuxer timeline (fixes #15814)
The "run-time" parameter was measured in terms of how long the input
thread had actually been playing, i.e. the real time since the thread
was created minus the real time spent in paused state.

This is a bit odd. It does not make much sense when transcoding, and
also seems rather counter-intuitive when playing a non-nominal rate.

This patch changes the run-time variable to match the demuxer time, as
does the start-time and the stop-time.