CDG: fix seek
The video output is messed up if we only do a stream_Seek(). If we want
to seeking forward, we should decode and discard pictures to keep the
decoder at the right state. For seeking backward, seeking to the
beginning and decode from there. Since the bitrate is very low and the
decoding process is really simple, there is no obvious delay. I have
tested this strategy on a dual core ARM board.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>