variables: voidify callback addition functions
In case of memory error, the functions will abort (table macros). Even
if it did not, callback deletion would later abort failing to find the
callback instead. (The API could be redesigned not to require memory
allocation if necessary.)

In case of nonexistent variable, log an error and fail safe as before.
This lead to deadlocks of misbehaviours as before - no changes here.