build: disable non-PIC assembly when building for Android/x86
This disables MMX/MMXEXT code for ffmpeg/postproc, and all assembly for
mpg123 and libass, if building for Android on x86 (32-bit).

It has performance implications but it avoids all TEXTRELs in the output,
which makes it possible to use W^X hardening techniques. The code can
probably be ported to be PIC-compatible in the future so that these can be
re-enabled, but even Gentoo Linux has no better solutions for this at the
moment.

Signed-off-by: Diego Elio Pettenò <flameeyes@flameeyes.eu>
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>