mediacodec: implement drain
When draining (pp_block == NULL), queue a dummy input buffer with a EOS flag,
and wait for the output thread to process all output frames.

This patch also add extra check when processing output buffers. Indeed, in some
case, the last output buffer with a EOS flag can be invalid.