demux: adaptative: fix unsupported format regression
Due to hls we we waiting for demuxer to start, which would
load the demuxer source (sub playlist then) and from that
file we would have the format and change the demux if
required.

The problem is that any unsupported format (like split subtitles)
is now being to be demuxed and gets probed and demuxed.