Make ALSA devices without description appear in Qt4 config.
If an /etc/asound.conf or $HOME/.asoundrc creates devices, those do not have a
description (a null pointer is returned). Relatedly, redefining the "default"
device results in it being listed without a description, thus disabling the
code that adds "default" WITH a description.
Possibly fixes other problems where ALSA devices with a NULL
description are disregarded. In Qt4 config, the NULL check is in
StringListConfigControl::finish().

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>