http: explicitly allow deflate & gzip encoding in live streams
This explicitly allows the server to return gzip or deflate
Content-Encoding.

This is not explicitly allowed for normal streams because
Content-Encoding breaks use of bytes ranges. Seeking offsets and size
would be in terms of the compressed stream. This would prevent VLC
demuxers from seeking or estimating duration.

The HTTP origin server can still return Content-Encoding in any case as
VLC does not explicitly forbid it (with Accept-Encoding: identity).
However in practice, servers will not perform opportunistic on-the-fly
compression.