Implemented a new edge detection video filter
The edge detection filter uses the Sobel operator to detect and highlight
edges in a frame. In order to successfully detect an edge, the image must
first be converted to gray scale (because Sobel is a gray scale operator) and
then have a slight Gaussian blur applied to it. We do this because Sobel is a
noisy operator and the Gaussian blur reduces this noise.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>