ZVBI: add option to select decoding Presentation Level
This commit adds an option to the ZVBI settings to select the Teletext
Presentation Level which is used for decoding. Thus one can switch to a lower
level in case of problems or can compare the rendering of a page at different
levels. The default value is Level 3.5, which already has been used each time
for decoding here.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>