direct3d: Fix green lines.
This patch reworks a bit the way we handle textures & surfaces.
The surface is the size of the decoder buffer (i_width/i_height)
The texture is the size of the visible picture (i_visible_*)
StretchRect is now copying only the valuable parts of the decoder
buffer (ie. it includes offsets & visible_*)
Fix #16597