mkv: added warning regarding undefined-behavior
The workaround currently present in the codebase is undefined-behavior
since the dynamic type of the object in question is not the one used in
the `static_cast`.

In short nothing guarantees that the vtable for the different types look
the same, and as such this is very dangerous.

We should probably supply a patch upstream if the bug is still present
in their code.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>