mkv: Introduced MKV_CHECKED_PTR_DECL
`MKV_IS_ID` is a macro to check whether the object hiding behind a pointer
has a certain dynamic type and it is being used all over the module.

There are however a lot of cases that is followed by a static_cast to the
appropriate type; `MKV_CHECKED_PTR_DECL` has been introduced to shorten
code that has such usage.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>