mkv: fixed C++03 restriction regarding IndexFinder
It somehow slipped my mind that C++03 disallows local types as
template-arguments, which makes one of the earlier commits invalid when
compiled as C++03.

This patch correctly moves the type in question to the global namespace
(inside an anonymous namespace to not pollute the global linkage scope)
so that everything is 100% legal C++03.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>