mkv: fix build failure when compiled with llvm
LLVM treats a typedef with forward-declaration inside a local struct as
a forward-declaration for a nested type of that struct, effectively
erroring on the implementation of the dispatchers.

This patch fixes that issue by forcing the name to referr to a type in
the sourrounding scope (by introducing a proper forward-declaration).

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>