mkv: removed duplicate case for S_VOBSUB
During the refactor to use the new StringDispatcher it seems the code
responsible for detecting the S_VOBSUB codec was added twice (but logic
only for the first entry).

Even though the first entry will take precedence (and hence make it
work), we should of course not have unused and unnecessary code in the
code base.

This patch removes the duplicate entry.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>