mkv: fixed faulty logic in matroska_Segment_c::ParseInfo
Steve "robux4" Lhomme noticed an error in the code-refactoring that
would create a new chapter_translation_c for every subelement found,
even though the found subelements should belong to the same
chapter_translation_c.

This patch fixes that by introducing a nested dispatch-handler for the
EbmlTypes belonging to a KaxChapterTranslate.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>