mkv: fixed broken locking mechanism
A temporary object (ie. an object without a name) is created in the
expression in which it is introduced, and destroyed immediately after.

This means that the previous code did not lock p_sys->lock_demuxer for
the remaining duration of the function "Demux", but instead _only_
during the expression in which the temporary "vlc_mutex_locker" was
constructed (ie. the locking and unlocking happened on that line).

This patch fixes the issue by giving the previously unnamed object a
name (binding its lifetime to the current scope).

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>