qt: make seek tooltip + actual seek less surprising
Previously seeking could be somewhat of a surprise if one relied on the
text present in the tooltip above the SeekSlider when hovering it with
your mouse.

Since the position of where we are going to end up was calculated
differently in different parts of the code, you could aim for "00:10:42"
but end up quite far from there (how far depends on the length of the
media you are playing).

This patch fixes the above by introducing two helper functions,
effectively making all calculations that are based on the X-coordinate
relative to the SeekSlider the same, and as such less surprising.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>