playlist/fetcher: fix vlc_interrupt assert
The fetcher thread is not attached, and is not joined (and can't be joined).
When playlist_fetcher_Delete() wait for the thread to finish (b_live == false),
we can't be sure that the thread really exited, and that the thread interrupt
TSD is destroyed. To fix this issue, unset the interrupt context manually from
the Thread just before it's terminated.

This problem was brought by 7f8d972a2b0ff62ef02a9d562a9a8299dd2eb985 and was
breaking some tests (non systematic).