mkv: removed unnecessary (and problematic) static data-member from EbmlTypeDispatcher
The previous static data-member really served no purpose, as such it has
been removed and been replaced by a more appropriate less-than operator
for EbmlProcessorEntry.

The static data-member caused issues when the module was compiled using msvc
(thanks for robux4 for finding the issue), which makes it even more important
for the patch to come into play.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>