mkv: cleanup of chapter_codec_cmds_c::AddCommand
This function was unnecessary complex, shortened the code by introducing
a helper variable that we can use for direct indexing instead of going
through a switch-statement.

In the future we might want to revert this change, but since we will
always have values in the range [0, 4) I do not see a problem with
shortening the function this way.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>