mkv: removed code dealing with KaxClusterSilentTrackNumber
This code has been removed since it does not aid us in any way (besides
cluttering up the implementation with a data-member that we do not
really care about).

In the future we should respect the relevant matroska element, but
currently there is no point for us to care about it given the demuxers
overall state.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>