mkv: removed unnecessary complexity in Ebml_parser.cpp
For starters it is ill-formed to have zero-sized arrays according to the
ISO C++ Standard, but most importantly the code is unnecessary complex
since there is no need for the array at all (given that the elements
inside of it has been commented out).

This commit also changes a comment which has not been accurate for a
long time, though it would be possible to simplify the code in question
-- it is hard to prove that the behavior will not change, and as such it
has been left as is.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>