input: always sanitize filenames from formatted meta-data
The input item name can contain special characters. It is not too hard
to imagine how this could be exploited.

path_sanitize() is inadequate as it cannot differentiate special
characters that are part of the trusted format string from special
characters that came from expanding the input item name. Instead,
filename_sanitize() must be used to eliminate all special characters.