Fix OSDEpgText size and opacity
This patch will fix the OSDEpg (triggered by the "position" hotkey.) not
showing text information because of transparent alpha value and invalid
font size.
The argument size of the vout_OSDEpgText function is already in pixels and
not relative, so p_style->i_font_size is the correct style option.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>