mkv: store priority tracks
When seeking, we should prioritize tracks depending on what type of ES
they are; this patch addes a data-member to matroska_segment_c that will
keep track of which tracks are more important than other.

It also adds logic to ComputePriorityTracks that will initialize this
data-member with the appropriate track-ids.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>