mkv: add cue position regardless of partial support or not
Even though there are currently some type of cues that we cannot handle,
we should add the aprox. location denoted by such cues as a seekpoint to
optimize seeking.

Adding the location with Seekpoint::DISABLED means that we will never
seek directly to the position, but it can be used as a hint when looking
for an appropriate keyframe during the preparation for seeking.

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>