contrib: don't unnecessary rebuild microdns
As a rule the content of a tarball doesn't have newer timestamps than the
tarball file itself. Thus after unpacking the directory microdns has an
older timestamp than the tarball file. Every time make runs, it thinks the
tarball has been updated and microdns must be rebuilt.

To avoid this, additionally touch the directory microdns after unpacking,
which is already included in a predefined command MOVE.

Signed-off-by: Thomas Guillem <thomas@gllm.fr>